refactor: 서비스단 컨트롤러로 이동
